pyproject.toml
[tool.poetry]
name = "python_base_project"
version = "0.0.8"
description = "A sample Python project"
authors = ["Bella Belgarokova <bella.belgarokova@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.26.0"
click = ">=8.0.2"
isort = "^5.13.2"
langchain = "^0.2.11"
langchain-openai = "^0.1.19"
langchain-groq = "^0.1.8"
langchain-community = "^0.2.10"
gitpython = "^3.1.43"
langchain-core = "^0.2.24"
python-dotenv = "^1.0.1"
pre-commit = "^3.8.0"
coverage = "^7.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

isort = "^5.13.2"

[tool.pytest.ini_options]
addopts = "--disable-warnings --cache-clear"
cache_dir = "/dev/null"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
